+++ tests/ll_queue_tests.c
@@ -14,7 +14,7 @@ bool empty_test(void) {
     }
     bool res = true;
 
-    res &= is_equal(ll_queue_get_count(queue_p), (size_t)0);
+    res &= is_equal(ll_queue_get_count(queue_p), 0UL);
     res &= is_true(ll_queue_is_empty(queue_p));
     res &= is_true(ll_queue_deinit(&queue_p));
 
@@ -34,11 +34,11 @@ bool one_element_test(void) {
     res &= is_equal(value, ll_queue_peek_int(queue_p));
 
     res &= is_false(ll_queue_is_empty(queue_p));
-    res &= is_equal(ll_queue_get_count(queue_p), (size_t)1);
+    res &= is_equal(ll_queue_get_count(queue_p), 1UL);
 
     res &= is_equal(value, ll_queue_dequeue_int(queue_p));
     res &= is_true(ll_queue_is_empty(queue_p));
-    res &= is_equal(ll_queue_get_count(queue_p), (size_t)0);
+    res &= is_equal(ll_queue_get_count(queue_p), 0UL);
 
     res &= is_true(ll_queue_deinit(&queue_p));
 
@@ -59,7 +59,7 @@ bool two_elements_test(void) {
     res &= is_true(ll_queue_enqueue_int(queue_p, value2));
     res &= is_equal(value1, ll_queue_peek_first_int(queue_p));
     res &= is_equal(value2, ll_queue_peek_last_int(queue_p));
-    res &= is_equal(ll_queue_get_count(queue_p), (size_t)2);
+    res &= is_equal(ll_queue_get_count(queue_p), 2UL);
     res &= is_true(ll_queue_deinit(&queue_p));
 
     return res;
