+++ tests/ll_stack_tests.c
@@ -15,7 +15,7 @@ bool empty_test(void) {
     }
     bool res = true;
 
-    res &= is_equal(ll_stack_get_count(stack_p), (size_t)0);
+    res &= is_equal(ll_stack_get_count(stack_p), 0UL);
     res &= is_true(ll_stack_is_empty(stack_p));
     res &= is_true(ll_stack_deinit(&stack_p));
 
@@ -34,11 +34,11 @@ bool one_element_test(void) {
     res &= is_true(ll_stack_push_int(stack_p, value));
     res &= is_equal(value, ll_stack_peek_int(stack_p));
     res &= is_false(ll_stack_is_empty(stack_p));
-    res &= is_equal(ll_stack_get_count(stack_p), (size_t)1);
+    res &= is_equal(ll_stack_get_count(stack_p), 1UL);
 
     res &= is_equal(value, ll_stack_pop_int(stack_p));
     res &= is_true(ll_stack_is_empty(stack_p));
-    res &= is_equal(ll_stack_get_count(stack_p), (size_t)0);
+    res &= is_equal(ll_stack_get_count(stack_p), 0UL);
 
     res &= is_true(ll_stack_deinit(&stack_p));
